moving the "long long" define to axPlugin, right before the Format checks. "should be fixed" now.
last couple of warnings are:

C:\axonlib\axonlib\plugins\test>..\compile test_gain_nogui.cpp

.\test_gain_nogui.cpp:38: warning: first argument of 'int main(VstIntPtr (*)(AEffect*, VstInt32, Vst
Int32, VstIntPtr, void*, float))' should be 'int'
.\test_gain_nogui.cpp:38: warning: 'int main(VstIntPtr (*)(AEffect*, VstInt32, VstInt32, VstIntPtr,
void*, float))' takes only zero or two arguments

+++ ax/core/axMath.h
@@ -558,7 +558,7 @@ inline float axPowf(const float x, const float y)
  * @param[in] n int
  * @return result float
  */
-inline float axPow(float x, unsigned int n)
+inline float axPow(float x, unsigned long n)
 {
   float res = 1;
   while (n > 0)
